+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [2.0] - 2020-12-14
 
 ### Added
 - Add default values as placeholders in admin fields.
@@ -61,7 +61,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Fork of original [WP Mail From](https://wordpress.org/plugins/wp-mailfrom/) plugin
 
-[Unreleased]: https://github.com/benhuson/wp-mailfrom/compare/1.2...HEAD
+[Unreleased]: https://github.com/benhuson/wp-mailfrom/compare/2.0...HEAD
+[2.0]: https://github.com/benhuson/wp-mailfrom/compare/1.2...2.0
 [1.2]: https://github.com/benhuson/wp-mailfrom/compare/1.1.1...1.2
 [1.1.1]: https://github.com/benhuson/wp-mailfrom/compare/1.1...1.1.1
 [1.1]: https://github.com/benhuson/wp-mailfrom/compare/1.0.2...1.1